moved utility class to its own file
