+++ Assets/MixedRealityToolkit-Tests/InputSystem/InputSystemTestUtilities.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c48e6f5d0f4d4cddbaa3f0301b9a7c72
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
