+++ Assets/MixedRealityToolkit-Tests/InputSystem/TestFixture_03_InputSystemTests.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Core.Definitions.Devices;
-using Microsoft.MixedReality.Toolkit.Core.Definitions.InputSystem;
 using Microsoft.MixedReality.Toolkit.Core.Interfaces.InputSystem;
 using Microsoft.MixedReality.Toolkit.Core.Services;
 using Microsoft.MixedReality.Toolkit.SDK.Input;
@@ -12,27 +10,6 @@ using UnityEngine.TestTools;
 
 namespace Microsoft.MixedReality.Toolkit.Tests.InputSystem
 {
-    public static class InputSystemTestUtilities
-    {
-        public static MixedRealityInputSystemProfile SetupInputSystemProfile()
-        {
-            // Create a Input System Profiles
-            var inputSystemProfile = ScriptableObject.CreateInstance<MixedRealityInputSystemProfile>();
-            inputSystemProfile.FocusProviderType = typeof(FocusProvider);
-            inputSystemProfile.InputActionsProfile = ScriptableObject.CreateInstance<MixedRealityInputActionsProfile>();
-            inputSystemProfile.InputActionRulesProfile = ScriptableObject.CreateInstance<MixedRealityInputActionRulesProfile>();
-            inputSystemProfile.PointerProfile = ScriptableObject.CreateInstance<MixedRealityPointerProfile>();
-            inputSystemProfile.PointerProfile.GazeProviderType = typeof(GazeProvider);
-            inputSystemProfile.GesturesProfile = ScriptableObject.CreateInstance<MixedRealityGesturesProfile>();
-            inputSystemProfile.SpeechCommandsProfile = ScriptableObject.CreateInstance<MixedRealitySpeechCommandsProfile>();
-            inputSystemProfile.ControllerVisualizationProfile = ScriptableObject.CreateInstance<MixedRealityControllerVisualizationProfile>();
-            inputSystemProfile.ControllerMappingProfile = ScriptableObject.CreateInstance<MixedRealityControllerMappingProfile>();
-
-            return inputSystemProfile;
-        }
-    }
-
-
     public class TestFixture_03_InputSystemTests
     {
         [Test]
